Components http-server & wrapper, server-input-interface-adapter added A bug needs to be fixed
Now the user interaction is through web server instead of buttons

The home node device will be placed in los of the gate node.
The home node will connect with AP
The user will interact withe the server on home node and send it command
The home node will convey this using esp-now to gate node
The server-input-interface-adaper uses the http-server to provide the
user-interaction interface.
The server could have directly implemented the user interface, but that
seems not neat and server own identity is hidden.
Now server is a basic http server wothout knowledge of what interface it
will serve
And the adapter component uses it to provide the interface
Right now a bug needs to be fixed. The send function msg_sent_handler of
home_node calls the inform command status memeber of user output interface
The method assigned pushes to queue. same queue is waited by http server
to respond.