Add remote component links and GitHub Actions workflow (#1)
* 1st commit : Working verified

First ever project with dependency inversion
Working verified.
Has two buttons to send open and close gate commmand to the gate node

* The gate node added as peer before discovery, not optimum, simple and working

* Command status printed

* Components http-server & wrapper, server-input-interface-adapter added
A bug needs to be fixed

Now the user interaction is through web server instead of buttons

The home node device will be placed in los of the gate node.
The home node will connect with AP
The user will interact withe the server on home node and send it command
The home node will convey this using esp-now to gate node
The server-input-interface-adaper uses the http-server to provide the
user-interaction interface.
The server could have directly implemented the user interface, but that
seems not neat and server own identity is hidden.
Now server is a basic http server wothout knowledge of what interface it
will serve
And the adapter component uses it to provide the interface
Right now a bug needs to be fixed. The send function msg_sent_handler of
home_node calls the inform command status memeber of user output interface
The method assigned pushes to queue. same queue is waited by http server
to respond.

* Web server used for user interaction instead of button, Wifi station smartconfig component added

The old simple wifi init replaced by wifi smart config component
It waits till connected  to AP, then gets the channel assigned
Then inits espnow with that channel supplied in config.
Earlier a static value of channel i.e channel 1 was used

* Wifi SmartConfig component added

* New componentshttp-server and WifiSmarconfig added

* Bugfix: The discovery completion callback requirement of the discovery component assigned to NULL before passing

Earlier it was left unassigned to NULL, and in the code it was crashing bcz being called while not being assigned
It was also passing the check of !=NULL bcz it was indeed not equal to NULL, it was dangling

* Bugfix: Now queue is emptied before acting on the web request

This way the response that comes in the queue is due to the request and not bcz of earlier

* The unused requirement of esp_timer component removed

It was earlier added because maybe the solution of queue filled with unused entries can be solved through it.

But then other solution of emptying the queue before acting on the request was used, so this esp_timer is redundant

* ESPNOW Component path is remote git link now

* New files added to .gitignore

* Workflow.yml and python script to download the build added

The yml builds on push and pull request
The python script requires the github token to access the private repo

* ESPIDF v%,4.2 specified

* syntax error correction

* Remote git path used instead of local

* removed dependency.lock

* only necessary files uploaded now instead of whole build folder

* Project level CMakeLists.txt also added for upload

* working

* build folder also cached, earier only installation cached

* idf.py will be replaced in next modificaion

* successfully download and monitor

* new elf file also put in build folder. idf_monitor specifies the custom build path

* no change just white space, wanted to test workflow

* Monitor now also working correctly

* no change just white space, wanted to test workflow

* no change just white space, wanted to test workflow